requirements.txt
longport==2.1.12
pymysql==1.1.1
python-dotenv==1.1.0